+++ Riel/3주차/baekjoon-2480.js
@@ -1,0 +1,5 @@
+const fs = require('fs');
+let input = fs.readFileSync('/dev/stdin').toString().split(' ');
+let a = parseInt(input[0]);
+let b = parseInt(input[1]);
+let c = parseInt(input[2]);
