+++ Riel/5월-4주차/baekjoon- copy.js
@@ -1,0 +1,4 @@
+const fs = require('fs');
+let input = fs.readFileSync('/dev/stdin').toString().split(' ');
+let a = parseInt(input[0]);
+let b = parseInt(input[1]);
